+++ src/CMakeLists.txt
@@ -54,7 +54,7 @@ add_subdirectory(pasi)
 add_subdirectory(poroelast)
 add_subdirectory(poroelast_scatra)
 add_subdirectory(porofluid_pressure_based)
-add_subdirectory(porofluidmultiphase_ele)
+add_subdirectory(porofluid_pressure_based_ele)
 add_subdirectory(poromultiphase)
 add_subdirectory(poromultiphase_scatra)
 add_subdirectory(red_airways)
