+++ src/mat/4C_mat_fluidporo_multiphase.cpp
@@ -11,7 +11,7 @@
 #include "4C_global_data.hpp"
 #include "4C_mat_fluidporo_singlephase.hpp"
 #include "4C_mat_par_bundle.hpp"
-#include "4C_porofluidmultiphase_ele_calc_utils.hpp"
+#include "4C_porofluid_pressure_based_ele_calc_utils.hpp"
 #include "4C_utils_enum.hpp"
 
 #include <Teuchos_SerialDenseSolver.hpp>
