+++ src/porofluid_pressure_based_ele/4C_porofluid_pressure_based_ele.cpp
@@ -5,7 +5,7 @@
 //
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#include "4C_porofluidmultiphase_ele.hpp"
+#include "4C_porofluid_pressure_based_ele.hpp"
 
 #include "4C_comm_pack_helpers.hpp"
 #include "4C_comm_utils_factory.hpp"
