+++ src/porofluid_pressure_based_ele/4C_porofluid_pressure_based_ele_boundary_factory.cpp
@@ -5,12 +5,12 @@
 //
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#include "4C_porofluidmultiphase_ele_boundary_factory.hpp"
+#include "4C_porofluid_pressure_based_ele_boundary_factory.hpp"
 
 #include "4C_fem_general_element.hpp"
 #include "4C_global_data.hpp"
-#include "4C_porofluidmultiphase_ele_boundary_calc.hpp"
-#include "4C_porofluidmultiphase_ele_interface.hpp"
+#include "4C_porofluid_pressure_based_ele_boundary_calc.hpp"
+#include "4C_porofluid_pressure_based_ele_interface.hpp"
 #include "4C_utils_enum.hpp"
 
 FOUR_C_NAMESPACE_OPEN
