+++ src/porofluid_pressure_based_ele/4C_porofluid_pressure_based_ele_boundary_factory.hpp
@@ -5,8 +5,8 @@
 //
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#ifndef FOUR_C_POROFLUIDMULTIPHASE_ELE_FACTORY_HPP
-#define FOUR_C_POROFLUIDMULTIPHASE_ELE_FACTORY_HPP
+#ifndef FOUR_C_POROFLUID_PRESSURE_BASED_ELE_BOUNDARY_FACTORY_HPP
+#define FOUR_C_POROFLUID_PRESSURE_BASED_ELE_BOUNDARY_FACTORY_HPP
 
 
 #include "4C_config.hpp"
@@ -22,31 +22,27 @@ namespace Discret
     // forward declaration
     class PoroFluidMultiPhaseEleInterface;
 
-    class PoroFluidMultiPhaseFactory
+    class PoroFluidMultiPhaseBoundaryFactory
     {
      public:
       //! ctor
-      PoroFluidMultiPhaseFactory() { return; }
+      PoroFluidMultiPhaseBoundaryFactory() { return; };
 
       //! dtor
-      virtual ~PoroFluidMultiPhaseFactory() = default;
+      virtual ~PoroFluidMultiPhaseBoundaryFactory() = default;
+
       //! ProvideImpl
       static PoroFluidMultiPhaseEleInterface* provide_impl(
-          Core::FE::CellType distype, const int numdofpernode, const std::string& disname);
+          const Core::Elements::Element* ele, const int numdofpernode, const std::string& disname);
 
      private:
-      //! define PoroFluidMultiPhaseEle instances dependent on problem
+      //! return instance of element evaluation class depending on implementation type
       template <Core::FE::CellType distype>
       static PoroFluidMultiPhaseEleInterface* define_problem_type(
           const int numdofpernode, const std::string& disname);
-
-
-    };  // end class PoroFluidMultiPhaseFactory
-
+    };  // class PoroFluidMultiPhaseBoundaryFactory
   }  // namespace Elements
-
 }  // namespace Discret
-
 
 
 FOUR_C_NAMESPACE_CLOSE
