+++ src/porofluid_pressure_based_ele/4C_porofluid_pressure_based_ele_calc_utils.cpp
@@ -5,7 +5,7 @@
 //
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#include "4C_porofluidmultiphase_ele_calc_utils.hpp"
+#include "4C_porofluid_pressure_based_ele_calc_utils.hpp"
 
 #include "4C_mat_fluidporo_multiphase.hpp"
 #include "4C_mat_fluidporo_multiphase_reactions.hpp"
