+++ src/porofluid_pressure_based_ele/4C_porofluid_pressure_based_ele_calc_utils.hpp
@@ -5,8 +5,8 @@
 //
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#ifndef FOUR_C_POROFLUIDMULTIPHASE_ELE_CALC_UTILS_HPP
-#define FOUR_C_POROFLUIDMULTIPHASE_ELE_CALC_UTILS_HPP
+#ifndef FOUR_C_POROFLUID_PRESSURE_BASED_ELE_CALC_UTILS_HPP
+#define FOUR_C_POROFLUID_PRESSURE_BASED_ELE_CALC_UTILS_HPP
 
 #include "4C_config.hpp"
