+++ src/porofluid_pressure_based_ele/4C_porofluid_pressure_based_ele_interface.hpp
@@ -5,8 +5,8 @@
 //
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#ifndef FOUR_C_POROFLUIDMULTIPHASE_ELE_INTERFACE_HPP
-#define FOUR_C_POROFLUIDMULTIPHASE_ELE_INTERFACE_HPP
+#ifndef FOUR_C_POROFLUID_PRESSURE_BASED_ELE_INTERFACE_HPP
+#define FOUR_C_POROFLUID_PRESSURE_BASED_ELE_INTERFACE_HPP
