+++ src/poromultiphase/4C_poromultiphase_utils.cpp
@@ -12,8 +12,8 @@
 #include "4C_global_data.hpp"
 #include "4C_inpar_bio.hpp"
 #include "4C_poroelast_utils.hpp"
+#include "4C_porofluid_pressure_based_ele.hpp"
 #include "4C_porofluid_pressure_based_utils.hpp"
-#include "4C_porofluidmultiphase_ele.hpp"
 #include "4C_poromultiphase_adapter.hpp"
 #include "4C_poromultiphase_monolithic_twoway.hpp"
 #include "4C_poromultiphase_partitioned_twoway.hpp"
