+++ src/poromultiphase/4C_poromultiphase_utils_clonestrategy.cpp
@@ -12,7 +12,7 @@
 #include "4C_mat_material_factory.hpp"
 #include "4C_mat_par_bundle.hpp"
 #include "4C_material_parameter_base.hpp"
-#include "4C_porofluidmultiphase_ele.hpp"
+#include "4C_porofluid_pressure_based_ele.hpp"
 
 FOUR_C_NAMESPACE_OPEN
