+++ src/poromultiphase_scatra/4C_poromultiphase_scatra_artery_coupling_nonconforming.cpp
@@ -15,8 +15,8 @@
 #include "4C_linalg_utils_sparse_algebra_manipulation.hpp"
 #include "4C_linalg_utils_sparse_algebra_print.hpp"
 #include "4C_mat_cnst_1d_art.hpp"
+#include "4C_porofluid_pressure_based_ele_parameter.hpp"
 #include "4C_porofluid_pressure_based_utils.hpp"
-#include "4C_porofluidmultiphase_ele_parameter.hpp"
 #include "4C_poromultiphase_scatra_artery_coupling_defines.hpp"
 #include "4C_poromultiphase_scatra_artery_coupling_pair.hpp"
 #include "4C_rebalance_binning_based.hpp"
