+++ src/poromultiphase_scatra/4C_poromultiphase_scatra_artery_coupling_pair.cpp
@@ -17,7 +17,7 @@
 #include "4C_global_data.hpp"
 #include "4C_mat_cnst_1d_art.hpp"
 #include "4C_mat_fluidporo_multiphase.hpp"
-#include "4C_porofluidmultiphase_ele_parameter.hpp"
+#include "4C_porofluid_pressure_based_ele_parameter.hpp"
 #include "4C_poromultiphase_scatra_artery_coupling_defines.hpp"
 #include "4C_scatra_ele_parameter_timint.hpp"
 #include "4C_utils_enum.hpp"
