+++ src/scatra_ele/4C_scatra_ele_calc_multiporo_reac.hpp
@@ -17,13 +17,13 @@
 #include "4C_mat_fluidporo_multiphase_reactions.hpp"
 #include "4C_mat_fluidporo_singlephase.hpp"
 #include "4C_mat_scatra_multiporo.hpp"
-#include "4C_porofluidmultiphase_ele_action.hpp"
-#include "4C_porofluidmultiphase_ele_calc.hpp"
-#include "4C_porofluidmultiphase_ele_calc_utils.hpp"
-#include "4C_porofluidmultiphase_ele_evaluator.hpp"
-#include "4C_porofluidmultiphase_ele_parameter.hpp"
-#include "4C_porofluidmultiphase_ele_phasemanager.hpp"
-#include "4C_porofluidmultiphase_ele_variablemanager.hpp"
+#include "4C_porofluid_pressure_based_ele_action.hpp"
+#include "4C_porofluid_pressure_based_ele_calc.hpp"
+#include "4C_porofluid_pressure_based_ele_calc_utils.hpp"
+#include "4C_porofluid_pressure_based_ele_evaluator.hpp"
+#include "4C_porofluid_pressure_based_ele_parameter.hpp"
+#include "4C_porofluid_pressure_based_ele_phasemanager.hpp"
+#include "4C_porofluid_pressure_based_ele_variablemanager.hpp"
 #include "4C_scatra_ele_calc_poro_reac.hpp"
 #include "4C_utils_enum.hpp"
 #include "4C_utils_function.hpp"
