+++ src/scatra_ele/CMakeLists.txt
@@ -16,7 +16,7 @@ set(_dependencies
     global_data
     inpar
     mat
-    porofluidmultiphase_ele
+    porofluid_pressure_based_ele
     )
 
 four_c_add_dependency(${AUTO_DEFINED_MODULE_NAME} ${_dependencies})
